move conf/parse to subpackage to resolve circular imports
