+++ pkg/conf/parse/diff.go
@@ -1,4 +1,4 @@
-package conf
+package parse
 
 import (
 	"fmt"
