+++ pkg/conf/parse/diff_test.go
@@ -1,4 +1,4 @@
-package conf
+package parse
 
 import (
 	"reflect"
